refactor DisableButtonsByPerspectiveListener
